Switch to regular integers for simpler code

+++ convert.c
@@ -3,9 +3,9 @@
 #define S(i, j) *(src + (i) + (j) * ldsrc)
 #define D(i, j) *(dst + (i) + (j) * lddst)
 
-void convert_double_to_float(double *src, uint64_t ldsrc, float *dst,
-                             uint64_t lddst, uint64_t m, uint64_t n) {
-    uint64_t i, j;
+void convert_double_to_float(double *src, int ldsrc, float *dst,
+                             int lddst, int m, int n) {
+    int i, j;
     for (i = 0; i < m; ++i) {
         for (j = 0; j < n; ++j) {
             D(i, j) = (float)S(i, j);
@@ -14,9 +14,9 @@ void convert_double_to_float(double *src, uint64_t ldsrc, float *dst,
     return;
 }
 
-void convert_float_to_double(float *src, uint64_t ldsrc, double *dst,
-                             uint64_t lddst, uint64_t m, uint64_t n) {
-    uint64_t i, j;
+void convert_float_to_double(float *src, int ldsrc, double *dst,
+                             int lddst, int m, int n) {
+    int i, j;
     for (i = 0; i < m; ++i) {
         for (j = 0; j < n; ++j) {
             D(i, j) = (double)S(i, j);
