+++ matgen.c
@@ -7,7 +7,7 @@ int64_t state = 1;  // Can be any odd number. Not thread safe.
 
 // Multiplicative Congruential Generator (MCG)
 uint32_t mcg_rand() {
-    const uint64_t MULTIPLIER = 14647171131086947261U;
+    const unsigned long long int MULTIPLIER = 14647171131086947261ULL;
     state *= MULTIPLIER;
     return state >> 32;
 }
@@ -16,9 +16,9 @@ uint32_t mcg_rand() {
 double mcg_rand_double() { return ((double)mcg_rand()) / UINT32_MAX - 0.5; }
 
 // Generate a row diagonally dominant square matrix A.
-void matgen(double *A, uint64_t lda, uint64_t m) {
+void matgen(double *A, int lda, int m) {
 
-    uint64_t i, j;
+    int i, j;
 
     double *diag = (double *)malloc(m * sizeof(double));
     memset(diag, 0, m * sizeof(double));
@@ -37,8 +37,8 @@ void matgen(double *A, uint64_t lda, uint64_t m) {
     free(diag);
 }
 
-void vecgen(double *v, uint64_t n) {
-    uint64_t i;
+void vecgen(double *v, int n) {
+    int i;
     for (i = 0; i < n; i++) {
         v[i] = mcg_rand_double();
     }
